refactor: draw the disk component with *iced*

+++ src/config.rs
@@ -65,7 +65,7 @@ pub struct DiskBarProps {
 
 #[derive(Deserialize)]
 pub struct DiskProps {
-  pub update_interval: u32,
+  pub update_interval: u64,
   pub name: String,
   pub device_path: String,
   pub mount_point: String,
